requirements.txt
channels==2.3.1 
channels-redis==2.4.1  
djangorestframework==3.10.3 
djangorestframework-simplejwt==4.4.0 
Pillow
Django==2.2.8
psycopg2-binary==2.8.4 
pytest-asyncio==0.10.0 
pytest-django==3.7.0